style: Give custom errors their own files
Some errors were inline and others were files. This cleans
up the main code and gets everything consistent into their
own error class files.
