+++ lib/definition.js
@@ -2,75 +2,26 @@
 
 const assert = require('assert')
 const has = require('./has.js')
-const MissingEnvError = require('./missing-env-error.js')
-const MutexError = require('./required-default-mutex-error.js')
 const typeOf = require('./type-of.js')
-
 const {
   TRUE_VALUES,
   FALSE_VALUES
 } = require('./boolean-values.js')
+const {
+  EnumError
+, ListError
+, MissingEnvError
+, RegExpError
+, RequiredDefaultMutexError
+} = require('./error-types/index.js')
+
 const VALID_BOOLEANS = new Set([
   ...Array.from(TRUE_VALUES)
 , ...Array.from(FALSE_VALUES)
 ])
+
 const EMPTY_STRING = ''
 
-class RegExpError extends Error {
-  constructor({
-    actual,
-    expected,
-    env
-  }) {
-    const post = `Expected "${actual}" to match ${expected}`
-    super(`Invalid ENV VAR (${env}) - ${post}`)
-    this.expected = expected
-    this.actual = actual
-    this.env = env
-    Error.captureStackTrace(this, this.constructor)
-  }
-
-  get name() {
-    return 'RegExpError'
-  }
-}
-
-class EnumError extends Error {
-  constructor({
-    actual,
-    expected,
-    env
-  }) {
-    const post = `Expected "${actual}" to be in [${expected}]`
-    super(`Invalid ENV VAR (${env}) - ${post}`)
-    this.expected = expected
-    this.actual = actual
-    this.env = env
-    Error.captureStackTrace(this, this.constructor)
-  }
-
-  get name() {
-    return 'EnumError'
-  }
-}
-
-class ListError extends Error {
-  constructor(opts) {
-    const {type, env, actual, input, original, reason} = opts
-    const extra = reason || `Value ${actual} is not of type ${type}`
-    super(`Invalid value in List ENV VAR ${env} - ${extra} (${input})`)
-    this.type = type
-    this.env = env
-    this.actual = String(actual)
-    this.original = original
-    this.input = input
-    Error.captureStackTrace(this, this.constructor)
-  }
-
-  get name() {
-    return 'ListError'
-  }
-}
 const SEP_TYPES = new Set([
   'string'
 , 'regexp'
@@ -227,7 +178,7 @@ module.exports = class Definition {
 
   _prevalidate() {
     if (this._required && this._default !== null) {
-      const err = new MutexError({
+      const err = new RequiredDefaultMutexError({
         type: this._type
       , name: this._env
       })
