+++ lib/error-types/index.js
@@ -1,0 +1,9 @@
+'use strict'
+
+module.exports = {
+  EnumError: require('./enum-error.js')
+, ListError: require('./list-error.js')
+, MissingEnvError: require('./missing-env-error.js')
+, RegExpError: require('./regexp-error.js')
+, RequiredDefaultMutexError: require('./required-default-mutex-error.js')
+}
