+++ lib/error-types/required-default-mutex-error.js
